+++ src/grammar/grammar.ohm-bundle.d.ts
@@ -22,7 +22,7 @@ export interface TactActionDict<T> extends ActionDict<T> {
   NativeFunction?: (this: NonterminalNode, arg0: NonterminalNode) => T;
   Type_optional?: (this: NonterminalNode, arg0: NonterminalNode, arg1: TerminalNode) => T;
   Type_required?: (this: NonterminalNode, arg0: NonterminalNode) => T;
-  Type_map?: (this: NonterminalNode, arg0: TerminalNode, arg1: TerminalNode, arg2: NonterminalNode, arg3: TerminalNode, arg4: NonterminalNode, arg5: TerminalNode) => T;
+  Type_map?: (this: NonterminalNode, arg0: TerminalNode, arg1: TerminalNode, arg2: NonterminalNode, arg3: IterationNode, arg4: IterationNode, arg5: TerminalNode, arg6: NonterminalNode, arg7: IterationNode, arg8: IterationNode, arg9: TerminalNode) => T;
   Type?: (this: NonterminalNode, arg0: NonterminalNode) => T;
   Field_default?: (this: NonterminalNode, arg0: NonterminalNode, arg1: TerminalNode, arg2: NonterminalNode, arg3: TerminalNode) => T;
   Field_defaultWithInit?: (this: NonterminalNode, arg0: NonterminalNode, arg1: TerminalNode, arg2: NonterminalNode, arg3: TerminalNode, arg4: NonterminalNode, arg5: TerminalNode) => T;